requirements.txt
autopep8==2.0.2
fastapi==0.99.1
uvicorn[standard]==0.22.0
SQLAlchemy==2.0.19
python-dotenv==1.0.0
PyMySQL==1.1.0
python-multipart==0.0.6